backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-storages[google]==1.14.2
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.18.0
python-docx==1.1.0
textract==1.6.5
officemath2latex==0.1.0
